requirements.txt
altair==5.3.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
openai==1.8.0
packaging==24.1
pandas==2.0.3
pillow==10.4.0
protobuf==5.27.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
py==1.11.0
pyarrow==17.0.0
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.9.1
Pygments==2.18.0
pyodbc==5.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
retry==0.9.2
rich==13.7.1
rpds-py==0.19.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sqlparse==0.3.0
sqlvalidator==0.0.2
streamlit==1.36.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
